+++ python/GafferCyclesTest/InteractiveCyclesRenderTest.py
@@ -49,6 +49,7 @@ import GafferCycles
 class InteractiveCyclesRenderTest( GafferSceneTest.InteractiveRenderTest ) :
 
 	interactiveRenderNodeClass = GafferCycles.InteractiveCyclesRender
+	renderer = "Cycles"
 
 	def testSVMRenderWithCPU( self ) :